% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr_function.R
\name{IDR}
\alias{IDR}
\title{A function to prepare for IDR.}
\usage{
IDR(MainPath, TestName, n_demo = 6, DIF = F)
}
\arguments{
\item{MainPath}{The folder where the response files are stored in}

\item{TestName}{Name of the test}

\item{n_demo}{number of demographic variables, usually it's 6.}

\item{DIF}{whether we are going to do DIF detection. The rule of thumb is each group should have more than 50 participants.
Thus, the default is F.}
}
\value{
Descriptive summary, dif results, CTT results and IRT results
}
\description{
Generate keycheck figures and DIF figures, perform DIF analysis, CTT analysis, and IRT analysis.
All the plots and csv files are stored in the folders. The results stored in txt files in jMetrik are stored as outputs of this function.
Reading is always analyzed before Listening.
}
\examples{
IDR(MainPath, TestName, n_demo=6)
}
