% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_function.R
\name{Scoring}
\alias{Scoring}
\title{A function to score the responses, compute percentage correct, and classify based on it.
Reading is always analyzed before Listening.}
\usage{
Scoring(MainPath, Language, TestName, AdminDate, ProfScale, ProfVersion_l,
  ProfVersion_r, n_demo = 6, PL_file = F, keychange_r = NULL,
  keychange_l = NULL)
}
\arguments{
\item{MainPath}{The folder where the response files are stored in}

\item{Language}{Testing language}

\item{TestName}{Name of the test}

\item{AdminDate}{Date of test administration}

\item{ProfScale}{Proficiency scale, "ACTFL" or "ILR"}

\item{ProfVersion_l}{The version of the proficiency level file downloaded from ACTesting for listening}

\item{ProfVersion_r}{The version of the proficiency level file downloaded from ACTesting for reading}

\item{n_demo}{number of demographic variables, usually it's 6.}

\item{PL_file}{logical. Default is FALSE, PL file get from Proficiency file getting from ACTesting. If TRUE, additional PL file is provided. Usually after IDR.
After IDR, just change the last column of PL file, including changing proficiency level and change to DNS.}

\item{keychange_r}{Default is NULL. If the reading key is changed from IDR, a list of item index and new keys should be here.
The first element is item index, the rest are the keys. Each item with key change is an object in the list.}

\item{keychange_l}{Default is NULL. If the listening key is changed from IDR, a list of item index and new keys should be here.}
}
\value{
Scores, proficiency level report, and simple score report
}
\description{
A function to score the responses, compute percentage correct, and classify based on it.
Reading is always analyzed before Listening.
}
\examples{
Scoring(MainPath, Language, TestName, AdminDate, ProfScale, ProfVersion_l, ProfVerson_r, n_demo, PL_file=F,
keychange_r=list(c(1,1,2), c(13,1,4)), keychange_l=NULL)
}
