% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating_function.R
\name{Tucker_equating}
\alias{Tucker_equating}
\title{A function to do the Tucker linear equating.}
\usage{
Tucker_equating(MainPath, XPath, YPath, XName, YName, n_demoX = 6,
  n_demoY = 6, DNS_X_r = NULL, DNS_X_l = NULL, DNS_Y_r = NULL,
  DNS_Y_l = NULL, keychange_X_r = NULL, keychange_X_l = NULL,
  keychange_Y_r = NULL, keychange_Y_l = NULL, equate_coef_l = list(c_2015
  = c(-2.282018, 1.052596), c_2014 = c(-1.848358, 1.014028), c_2013 =
  c(2.308507, 0.9523797), c_2012 = c(3.089623, 0.9053153)),
  equate_coef_r = list(c_2015 = c(-2.978765, 1.013981), c_2014 = c(5.188769,
  0.8741792), c_2013 = c(1.171173, 0.9723494), c_2012 = c(-3.480369, 1.05726)))
}
\arguments{
\item{MainPath}{The folder where the response files are stored in}

\item{XName}{Name of the X test (without "_Listening" or "_Reading")}

\item{n_demoX}{Number of demographic variables in X. Default value is 6.}

\item{n_demoY}{Number of demographic variables in Y.}

\item{DNS_X_r}{Vector of not score items in X reading test. Default is NULL.}

\item{DNS_X_l}{Vector of not score items in X listening test.}

\item{DNS_Y_r}{Vector of not score items in Y reading test.}

\item{DNS_Y_l}{Vector of not score items in Y listening test.}

\item{keychange_X_r}{List of key change for X reading test. Default is NULL.
The first element is the item index, the rest are the keys.}

\item{keychange_X_l}{List of key change for X listening test.}

\item{keychange_Y_r}{List of key change for Y reading test.}

\item{keychange_Y_l}{List of key change for Y listening test.}

\item{equate_coef_l}{List of Tucker linear equating coefficience for Listening.
First element is intercept, second is slope.
e.g. c_2015 equates 2016 test score to 2015. When using this function, just fill in proper equating coefficience.}

\item{equate_coef_r}{List of Tucker linear equating coefficience for Reading.}

\item{Xpath}{Path of the old response file is stored in}

\item{Ypath}{Path of the new response file is stored in}

\item{XName}{Name of the Y test}
}
\value{
Stats for both forms of test and the equating table.
}
\description{
It produces ItemStats file (for both old test and new test) and equating table.
Plots of liking item score and total score distribution are also developed.
}
\examples{
Tucker_equating(MainPath, XPath, YPath, XName, YName, n_demoX=6, n_demoY=6,
 DNS_X_r=NULL, DNS_X_l=c(1,5), DNS_Y_r=NULL, DNS_Y_l=NULL,
 keychange_X_r=list(c(2,1,4),c(13, 2)), keychange_X_l=NULL, keychange_Y_r=NULL, kechange_Y_l=NULL,
 equate_coef_l=list(c_2015=c(-2.282018,1.052596),
                c_2014=c(-1.848358,1.014028),
                c_2013=c(2.308507,0.9523797),
                c_2012=c(3.089623,0.9053153)),
 equate_coef_r=list(c_2015=c(-2.978765,1.013981),
                c_2014=c(5.188769,0.8741792),
                c_2013=c(1.171173,0.9723494),
                c_2012=c(-3.480369,1.05726)))
}
